Some refactorings

+++ src/IISExpress.ts
@@ -8,55 +8,55 @@ import { v4 as uuidv4 } from 'uuid';
 import * as iconv from 'iconv-lite';
 
 export interface IExpressArguments {
-	path?: string;
-	port?: number;
-	clr?: settings.clrVersion;
-	protocol?: settings.protocolType;
+	path: string;
+	port: number;
+	clr: settings.clrVersion;
+	protocol: settings.protocolType;
 }
 
 export class IISExpress {
 	private _iisProcess!: process.ChildProcessWithoutNullStreams;
 	private _iisPath: string;
 	private _iisAppCmdPath: string;
-	private _args: IExpressArguments;
+	private _args!: IExpressArguments;
 	private _output!: vscode.OutputChannel | null;
 	private _statusbar!: vscode.StatusBarItem;
 	private _statusMessage!: string;
 
-	constructor(iisPath: string, appCmdPath: string, args: IExpressArguments){
+	constructor(iisPath: string, appCmdPath: string){
 		this._iisPath = iisPath;
 		this._iisAppCmdPath = appCmdPath;
-		this._args = args;
 	}
 
-	public startWebsite(options?: settings.Isettings) {
+	public startWebsite(options: settings.Isettings) {
 
 		// Verify process not already running, so if we have a PID (process ID) it's running
-		if(this._iisProcess !== undefined){
+		if(this._iisProcess !== undefined && this._iisProcess.killed === false){
 			// Display error message that it's already running
-			vscode.window.showErrorMessage('IISExpress is already running');
+			vscode.window.showErrorMessage('IIS Express is already running',{modal: true});
 
 			// Stop the method/function from running
 			return;
 		}
 
-
 		if(!options){
-			vscode.window.showErrorMessage('No options found');
+			vscode.window.showErrorMessage('No options found in .vscode/iisexpress.json',{modal: true});
 			return;
 		}
 
-        // Get IIS Port Number from config file
-        this._args.port = options.port;
+		this._args = {
+			// Get IIS Port Number from config file
+			port: options.port,
 
-		// Folder to run as the arg
-		this._args.path = options.path ? options.path : vscode.workspace.rootPath;
+			// Folder to run as the arg
+			path: options.path ? options.path : <string>vscode.workspace.rootPath,
 
-		// CLR version, yes there are still people on 3.5 & default back to v4 if not set
-		this._args.clr = options.clr ? options.clr : settings.clrVersion.v40;
+			// CLR version, yes there are still people on 3.5 & default back to v4 if not set
+			clr: options.clr ? options.clr : settings.clrVersion.v40,
 
-		// If no protocol set fallback to http as opposed to https
-		this._args.protocol = options.protocol ? options.protocol : settings.protocolType.http;
+			// If no protocol set fallback to http as opposed to https
+			protocol: options.protocol ? options.protocol : settings.protocolType.http
+		};
 
 		// Create output channel & show it
 		this._output = this._output || vscode.window.createOutputChannel('IIS Express');
@@ -70,7 +70,7 @@ export class IISExpress {
 		// Then display an error & stop process
 		if(this._args.protocol === settings.protocolType.https && (this._args.port >= 44300 && this._args.port <=44399) === false){
 			// Using HTTPS but not using a port within the range that supports SSL
-			vscode.window.showErrorMessage('When using HTTPS you need to use ports 44300 - 44399 in .vscode/iisexpress.json');
+			vscode.window.showErrorMessage('When using HTTPS you need to use ports 44300 - 44399 in .vscode/iisexpress.json',{modal: true});
 
 			// Stop the method/function from running
 			return;
@@ -152,6 +152,13 @@ export class IISExpress {
 	}
 
 	public stopWebsite(){
+		// Ensure process is not null & is running
+		if(this._iisProcess === undefined || this._iisProcess.killed === true){
+			// Display error message
+			vscode.window.showErrorMessage('IIS Express is not running', {modal:true});
+			return;
+		}
+
 		// Kill the process - which will also hook into the exit event to remove the config entry
 		this._iisProcess.kill('SIGINT');
 
@@ -167,7 +174,7 @@ export class IISExpress {
 
 	}
 
-	public openWebsite(options?: settings.Isettings){
+	public openWebsite(options: settings.Isettings){
 
 		let browserConfig = vscode.workspace.getConfiguration().get<string>("iisexpress.openInBrowser", "default");
 		let browser:string = browserConfig.toLocaleLowerCase() === "default" ? "" : browserConfig;
@@ -184,7 +191,7 @@ export class IISExpress {
 		}
 	}
 
-	public restartSite(options? : settings.Isettings){
+	public restartSite(options : settings.Isettings){
 		// If we do not have an iisProcess/website running
 		if(!this._iisProcess){
 			// Then just do a start site...
